refactor: Provide separate `userSettings` source for both extensions
