+++ node/packages/aws-lambda-otel-extension/opt/otel-extension/external/user-settings.js
