+++ node/packages/aws-lambda-otel-extension/opt/otel-extension/internal/user-settings.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const settingsJsonString = process.env.SLS_OTEL_USER_SETTINGS;
+
+if (!settingsJsonString) return;
+
+try {
+  module.exports = JSON.parse(settingsJsonString);
+} catch (error) {
+  process.stdout.write(`Resolution of user settings failed with: ${error.message}`);
+}
